removed print and moved util func
